refactor: CCP-4535 move group list container
