+++ frontend/src/__tests__/components/group/GroupListCard.test.ts
@@ -2,10 +2,11 @@ import { mount } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 
 import { makeGroup } from '@/__tests__/__factories__'
+
 import GroupListCard from '@/components/group/GroupListCard.vue'
 import { type Group } from '@/models/group.model'
 
-const mountComponent = (props: { group: Group; isAdmin: boolean }) =>
+const mountComponent = (props: { group: Group }) =>
   mount(GroupListCard, {
     props,
     global: {
@@ -15,7 +16,6 @@ const mountComponent = (props: { group: Group; isAdmin: boolean }) =>
           emits: ['click'],
         },
         'v-card-title': { template: '<div><slot /></div>' },
-        'v-card-text': { template: '<div><slot /></div>' },
       },
     },
   })
@@ -24,45 +24,15 @@ describe('GroupListCard.vue', () => {
   describe('group info', () => {
     it('renders the group name', () => {
       const group = makeGroup({ name: 'Administrators' })
-      const wrapper = mountComponent({ group, isAdmin: false })
+      const wrapper = mountComponent({ group })
 
       expect(wrapper.text()).toContain('Administrators')
-    })
-
-    it('renders the created date', () => {
-      const group = makeGroup({ createdDate: '2026-03-15' })
-      const wrapper = mountComponent({ group, isAdmin: false })
-
-      expect(wrapper.text()).toContain('2026-03-15')
-    })
-
-    it('renders the created by', () => {
-      const group = makeGroup({ createdBy: 'jsmith' })
-      const wrapper = mountComponent({ group, isAdmin: false })
-
-      expect(wrapper.text()).toContain('jsmith')
-    })
-  })
-
-  describe('admin state', () => {
-    it('renders Edit Group when isAdmin is true', () => {
-      const wrapper = mountComponent({ group: makeGroup(), isAdmin: true })
-
-      expect(wrapper.text()).toContain('Edit Group')
-      expect(wrapper.text()).not.toContain('View Group')
-    })
-
-    it('renders View Group when isAdmin is false', () => {
-      const wrapper = mountComponent({ group: makeGroup(), isAdmin: false })
-
-      expect(wrapper.text()).toContain('View Group')
-      expect(wrapper.text()).not.toContain('Edit Group')
     })
   })
 
   describe('click', () => {
     it('emits click when the card is clicked', async () => {
-      const wrapper = mountComponent({ group: makeGroup(), isAdmin: false })
+      const wrapper = mountComponent({ group: makeGroup() })
 
       await wrapper.find('div').trigger('click')
